scripts/requirements.txt
supabase>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
